+++ tests/packerlicious/test_builder_vmware.py
@@ -1,0 +1,23 @@
+import pytest
+
+import packerlicious.builder as builder
+
+
+class TestVMwareIsoBuilder(object):
+
+    def test_required_fields_missing(self):
+        b = builder.VMwareIso()
+
+        with pytest.raises(ValueError) as excinfo:
+            b.to_dict()
+        assert 'required' in str(excinfo.value)
+
+
+class TestVMwareVmxBuilder(object):
+
+    def test_required_fields_missing(self):
+        b = builder.VMwareVmx()
+
+        with pytest.raises(ValueError) as excinfo:
+            b.to_dict()
+        assert 'required' in str(excinfo.value)